+++ shop/cascade/processbar.py
@@ -10,7 +10,7 @@ from django.forms.fields import IntegerField
 from django.template.loader import select_template
 from cms.plugin_pool import plugin_pool
 from cmsplugin_cascade.forms import ManageChildrenFormMixin
-from cmsplugin_cascade.fields import PartialFormField
+from cmsplugin_cascade.fields import GlossaryField
 from cmsplugin_cascade.link.forms import TextLinkFormMixin
 from cmsplugin_cascade.link.plugin_base import LinkElementMixin
 from cmsplugin_cascade.widgets import NumberInputWidget
@@ -70,11 +70,9 @@ class ProcessStepPlugin(TransparentMixin, ShopPluginBase):
     allow_children = True
     alien_child_classes = True
     render_template = 'cascade/generic/wrapper.html'
-    glossary_fields = (
-        PartialFormField('step_title',
-            widgets.TextInput(attrs={'size': 150}),
-            label=_("Step Title")
-        ),
+    step_title = GlossaryField(
+        widgets.TextInput(attrs={'size': 50}),
+        label=_("Step Title")
     )
 
     @classmethod
@@ -91,8 +89,7 @@ plugin_pool.register_plugin(ProcessStepPlugin)
 
 
 class ProcessNextStepForm(TextLinkFormMixin, ModelForm):
-    link_content = CharField(required=False, label=_("Button Content"),
-        widget=widgets.TextInput(attrs={'id': 'id_name'}))
+    link_content = CharField(required=False, label=_("Button Content"), widget=widgets.TextInput())
 
     def __init__(self, raw_data=None, *args, **kwargs):
         instance = kwargs.get('instance')
@@ -106,8 +103,10 @@ class ProcessNextStepPlugin(BootstrapButtonMixin, ShopPluginBase):
     name = _("Next Step Button")
     parent_classes = ('ProcessStepPlugin',)
     form = ProcessNextStepForm
-    fields = ('link_content', 'glossary',)
     model_mixins = (LinkElementMixin,)
+    fields = ('link_content', 'glossary')
+    glossary_field_order = ('button_type', 'button_size', 'button_options', 'quick_float',
+                            'icon_left', 'icon_right')
 
     class Media:
         css = {'all': ('cascade/css/admin/bootstrap.min.css', 'cascade/css/admin/bootstrap-theme.min.css',)}