+++ shop/context_processors.py
@@ -4,6 +4,7 @@ from __future__ import unicode_literals
 from django import __version__
 from django.conf import settings
 from shop.models.customer import CustomerModel
+from shop import settings as shop_settings
 
 
 def customer(request):
@@ -31,4 +32,14 @@ def version(request):
     """
     return {
         'DJANGO_VERSION': __version__,
+    }
+
+
+def ng_model_options(request):
+    """
+    Add ng-model-options to the context, since these settings must be configurable
+    """
+    return {
+        'EDITCART_NG_MODEL_OPTIONS': shop_settings.EDITCART_NG_MODEL_OPTIONS,
+        'ADD2CART_NG_MODEL_OPTIONS': shop_settings.ADD2CART_NG_MODEL_OPTIONS,
     }