+++ shop/models/defaults/order.py
@@ -37,7 +37,12 @@ class Order(order.BaseOrder):
         return self.get_number()
 
     def get_number(self):
-        return '{0}-{1}'.format(str(self.number)[:4], str(self.number)[4:])
+        return str(self.number)[:4] + '-' + str(self.number)[4:]
+
+    @classmethod
+    def resolve_number(cls, number):
+        number = number[:4] + number[5:]
+        return dict(number=number)
 
     def populate_from_cart(self, cart, request):
         self.shipping_address_text = cart.shipping_address.as_text()