+++ shop/urls/order.py
@@ -6,5 +6,5 @@ from shop.views.order import OrderView
 
 urlpatterns = [
     url(r'^$', OrderView.as_view()),
-    url(r'^(?P<pk>\d+)$', OrderView.as_view(many=False)),
+    url(r'^(?P<slug>[\w-]+)/?$', OrderView.as_view(many=False)),
 ]