+++ shop/urls/rest_api.py
@@ -12,6 +12,6 @@ router.register(r'watch', WatchViewSet, base_name='watch')
 router.register(r'checkout', CheckoutViewSet, base_name='checkout')
 
 urlpatterns = [
-    url(r'^select_product$', ProductSelectView.as_view(), name='select-product'),
+    url(r'^select_product/?$', ProductSelectView.as_view(), name='select-product'),
     url(r'^', include(router.urls)),
 ]